Fix eval

+++ src/evaluation/calcFID.py
@@ -28,10 +28,10 @@ def get_batch(dataset, loc, batch_size):
     batch_image = np.zeros((bs, config.n_pet_channels, config.pet_image_dim, config.pet_image_dim))
     for i, p in enumerate(image_paths):
         batch_image[i] = load_image(p)
-        
+    
     batch_image = batch_image.reshape(bs * config.n_pet_channels, 1, config.pet_image_dim, config.pet_image_dim)
     batch_image = batch_image.repeat(3, axis=1)
-    batch_image = transform(torch.from_numpy(batch_image)).to(device)
+    batch_image = transform(torch.from_numpy(batch_image)).float().to(device)
     return batch_image
 
 def get_inception_features(model, dataset):
